requirements.txt
six==1.13.0
numpy==1.17.4
Pillow==7.1.2
tensorflow==2.1.0
